\name{afapi-package}
\alias{afapi-package}
\alias{afapi}
\docType{package}
\title{
Access the appFigures web API.
}
\description{
Access the appFigures web API.
}
\details{
\tabular{ll}{
Package: \tab afapi\cr
Type: \tab Package\cr
Version: \tab 0.0.0.9000\cr
Date: \tab 2015-04-30\cr
License: \tab GPL (>= 2) \cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
