% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getEvents.R
\name{createEvent}
\alias{createEvent}
\title{Create a new event.}
\usage{
createEvent(caption, date, details, products, curlHandle, verbose = FALSE)
}
\arguments{
\item{caption}{Character. A short caption summarizing the event.}

\item{date}{Character with "yyyy-mm-dd" format.}

\item{details}{Character. Any additional notes or details
the user wants to include describing the event.}

\item{products}{A character or numeric vector of product ids
that the user wants to associate with the new event.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}
}
\description{
\code{createEvent} allows the user to add a new event.
}
\seealso{
Official documentation:
\url{http://docs.appfigures.com/api/reference/v2/events}.
}

