% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getArchive.R
\name{getArchivedReport}
\alias{getArchivedReport}
\title{Retrieve archived reports for each of the linked accounts.}
\usage{
getArchivedReport(type = c("all", "daily", "weekly", "monthly", "monthlyfree",
  "finance", "payment"), date, curlHandle, verbose = FALSE, orgJSON = FALSE)
}
\arguments{
\item{type}{Character. Choose between \code{"all", "daily",
"weekly", "monthly", "monthlyfree", "finance", or "payment"}.
Defaults to \code{"all"}.}

\item{date}{Character. Only return reports with a timestamp on
this date. Defaults to all dates.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}

\item{orgJSON}{Logical. Should the JSON string be returned
without being converted to R objects? Defaults to \code{FALSE}.}
}
\value{
A dataframe containing the requested reports.
}
\description{
\code{getArchivedReport} provides access to archived reports
from the appFigures web API
}
\seealso{
Official documentation:
\url{http://docs.appfigures.com/archive}.
}

