% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getExternalAccounts.R
\name{getExternalAccounts}
\alias{getExternalAccounts}
\title{Retrieve external account details.}
\usage{
getExternalAccounts(id = NULL, curlHandle, verbose = FALSE,
  orgJSON = FALSE)
}
\arguments{
\item{id}{Character or numeric external account id. The
default \code{NULL} value will return details on all accounts.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}

\item{orgJSON}{Logical. Should the JSON string be returned
without being converted to R objects? Defaults to \code{FALSE}.}
}
\value{
A data frame.
}
\description{
\code{getExternalAccounts} provides details of a user's
external accounts.
}
\seealso{
Official documentation:
\url{http://docs.appfigures.com/api/reference/v2/external-accounts}.
}

