% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatured.R
\name{getFeaturedReport}
\alias{getFeaturedReport}
\title{Retrieve full featured reports.}
\usage{
getFeaturedReport(product_id, end_date, start_date, country, curlHandle,
  verbose = FALSE, orgJSON = FALSE)
}
\arguments{
\item{product_id}{Numeric. A single appFigures-assigned
product id.}

\item{end_date}{Character string or date object. Date of last ranks
to be reported. Defaults to today. See Details.}

\item{start_date}{Character string or date object. Date of first
ranks to be reported. Defaults to the last 31 days. See Details.}

\item{country}{One or more country iso codes. Defaults to all
countries. See Details.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}

\item{orgJSON}{Logical. Should the JSON string be returned
without being converted to R objects? Defaults to \code{FALSE}.}
}
\value{
A single R dataframe.
}
\description{
\code{getFeaturedReport} provides access to detailed featured
data from the appFigures web API
}
\details{
For \code{start_date} and \code{end_date}, if the
supplied argument can be interepreted as a date or POSIX
object, any hour, minute, or second is ignored. If a string
is supplied, it should have the format: 'yyyy-MM-dd'. The largest
date range for a feature report is limited to 31 days. Anything
larger will return an error.


For \code{country} arguments, a complete list of supported
languages can be found using \code{\link{getStoreData}}.
}
\seealso{
Official documentation: 
\url{http://docs.appfigures.com/api/reference/v2/featured}.
}
