% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArchive.R
\name{getLatestReport}
\alias{getLatestReport}
\title{Retrieve latest report for each of the linked accounts.}
\usage{
getLatestReport(type = c("all", "daily", "weekly", "monthly", "yearly",
  "finance", "payment"), curlHandle, verbose = FALSE, orgJSON = FALSE)
}
\arguments{
\item{type}{Character. Choose between \code{"all", "daily",
"weekly", "monthly", "yearly", "finance", "payment"}. Defaults
to \code{"all"}.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}

\item{orgJSON}{Logical. Should the JSON string be returned
without being converted to R objects? Defaults to \code{FALSE}.}
}
\value{
A dataframe containing the requested reports.
}
\description{
\code{getLatestReport} provides access to the latest reports
from the appFigures web API
}
\seealso{
Official documentation: 
\url{http://docs.appfigures.com/archive}.
}
