% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProducts.R
\name{getProducts}
\alias{getProducts}
\title{Retrieve products by store id or by appFigures-assigned id.}
\usage{
getProducts(id, store = c("all", "apple", "google", "amazon", "windows"),
  curlHandle, verbose = FALSE, orgJSON = FALSE)
}
\arguments{
\item{id}{Either the string value \code{"mine"} or the id of a
specific product.}

\item{store}{A character string to filter products by store.
Defaults to "all" stores.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}

\item{orgJSON}{Logical. Should the JSON string be returned
without being converted to R objects? Defaults to \code{FALSE}.}
}
\value{
A data frame containing all publicly available fields.
Fields containing private and meta data will be added later.
}
\description{
\code{getProducts} requests product specific data from the
appFigures web API.
}
\details{
Products represent a trackable item in the appFigures
system. Products can be an app as well as an in-app purchase
or book. The user can request data on all their products by
setting the id variable to \code{"mine"}. Data for a single
product can be requested by either its store specific id or
its appFigures-assigned id.
}
\seealso{
Official documentation:
\url{http://docs.appfigures.com/products}.
}

