% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getArchive.R
\name{getRawReport}
\alias{getRawReport}
\title{Retrieve the raw report data.}
\usage{
getRawReport(id, curlHandle, verbose = FALSE)
}
\arguments{
\item{id}{Numeric. The id of requested report.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}
}
\value{
A character string containing the raw report.
}
\description{
\code{getRawReport} provides access to the raw report data
from the appFigures web API
}
\seealso{
Official documentation:
\url{http://docs.appfigures.com/archive}.
}

