% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getReviews.R
\name{getReviews}
\alias{getReviews}
\title{Retrieve review reports.}
\usage{
getReviews(product_ids, q = NULL, page = 1, count = 25, lang = NULL,
  author = NULL, country, version, stars, sort = c("date", "country",
  "stars"), start_date, end_date, curlHandle, verbose = FALSE,
  orgJSON = FALSE)
}
\arguments{
\item{product_ids}{Numeric. A vector of appFigures-assigned
product ids. Defaults to all products in the account.}

\item{q}{Character. Only return reviews matching this character
string.}

\item{page}{Integer. Page number. Defaults to one.}

\item{count}{Integer. The number of reviews to return. Any number
between 1 and 500 is valid. Defaults to 25.}

\item{lang}{Character. Short language code to translate reviews
into. See Details.}

\item{author}{Character. Name of an author to filter reviews.}

\item{country}{One or more country iso codes. Defaults to all
countries. See Details.}

\item{stars}{Integer. Values of one to five to filter reviews on.
Defaults to reviews with any number of stars.}

\item{sort}{Character. How to sort the resulting reviews. Choose
from \code{"date", "country", or "stars"}. Defaults to \code{"date"}.}

\item{start_date}{Character string or date object. Date of first
ranks to be reported. Defaults to the last 31 days. See Details.}

\item{end_date}{Character string or date object. Date of last ranks
to be reported. Defaults to today. See Details.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}

\item{orgJSON}{Logical. Should the JSON string be returned
without being converted to R objects? Defaults to \code{FALSE}.}

\item{versions}{Character. One of more version numbers of an app to
filter reviews on.}
}
\value{
A data frame containing the requested review report.
}
\description{
\code{getReviews} provides access to detailed review data
from the appFigures web API
}
\details{
For \code{start_date} and \code{end_date}, if the
supplied argument can be interepreted as a date or POSIX
object, any hour, minute, or second is ignored. If a string
is supplied, it should have the format: 'yyyy-MM-dd'.

Regarding the \code{lang} and \code{country} arguments, a
complete list of supported languages can be found using
\code{\link{getStoreData}}.

Also, since the results are paginated, the returned data frame
has an attached attribute called 'header', which provides data
on the total number of reviews available, the current page number,
and the total number of pages (given 'count' = xx).d
}
\seealso{
Official documentation:
\url{http://docs.appfigures.com/api/reference/v2/reviews}.
}

