% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getStoreData.R
\name{getStoreData}
\alias{getStoreData}
\title{Retrieve general metadata used in other API requests.}
\usage{
getStoreData(tables = c("categories", "countries", "currencies", "languages",
  "stores"), curlHandle, verbose = FALSE, orgJSON = FALSE)
}
\arguments{
\item{tables}{Character. One of five choices: "categories",
"countries", "currencies", "languages", "stores". See Details.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}

\item{orgJSON}{Logical. Should the JSON string be returned
without being converted to R objects? Defaults to \code{FALSE}.}
}
\value{
A data frame containing the appropriate table, or if
\code{tables = "stores"}, a list of five data frames.
}
\description{
\code{getStoreData} can request data on store categories, country
abbreviations, supported languages for review translation, and
supported currencies from the appFigures web API.
}
\details{
For \code{tables} set to \code{"categories", "countries",
"currencies", or "languages"}, a single data frame is returned
containing the relevant metadata available across all stores. If
\code{tables} is set to \code{"stores"}, a list of five data frames
is returned. The first four data frames in the list contains all
metadata relevant for a particular store (e.g. apple, google_play,
ect.). The last data frame summarizes all the supported ad
networks.
}
\seealso{
Official documentation:
\url{http://docs.appfigures.com/api/reference/v2/data}.
}

