% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsage.R
\name{getUsage}
\alias{getUsage}
\title{Retrieve account usage details.}
\usage{
getUsage(curlHandle, verbose = FALSE, orgJSON = FALSE)
}
\arguments{
\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}
}
\value{
A data frame
}
\description{
\code{createEvent} provides a detailed breakdown of the past 30
days worth of usage as well as a summary of total usage.
}
\seealso{
Official documentation: 
\url{http://docs.appfigures.com/usage}.
}
