% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArchive.R
\name{parseArchiveReport}
\alias{parseArchiveReport}
\title{Map JSON string to an R data frame.}
\usage{
parseArchiveReport(jsonText)
}
\description{
\code{parseArchiveReport} parses the JSON returned by a
featured request made to the appFigures web API.
}
