% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimates.R
\name{parseEstimates}
\alias{parseEstimates}
\title{Map JSON string to an R data frame.}
\usage{
parseEstimates(jsonText)
}
\description{
\code{parseEstimates} parses the JSON returned by an app rank request made to the
 appFigures web API.
}
