% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getEvents.R
\name{parseEvents}
\alias{parseEvents}
\title{Map JSON string to an R data frame.}
\usage{
parseEvents(jsonText)
}
\description{
\code{parseEvents} parses the JSON returned by a
featured request made to the appFigures web API.
}

