% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExternalAccounts.R
\name{parseExternalAccounts}
\alias{parseExternalAccounts}
\title{Map JSON string to an R data frame.}
\usage{
parseExternalAccounts(jsonText, id)
}
\description{
\code{parseExternalAccounts} parses the JSON returned by an
external account request made to the appFigures web API.
}
