% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFeatured.R
\name{parseFeaturedCounts}
\alias{parseFeaturedCounts}
\title{Map JSON string to an R dataframe.}
\usage{
parseFeaturedCounts(jsonText)
}
\description{
\code{parseFeaturedCounts} parses the JSON returned by a
featured count request made to the appFigures web API.
}

