% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatured.R
\name{parseFeaturedHistory}
\alias{parseFeaturedHistory}
\title{Map JSON string to an R dataframe.}
\usage{
parseFeaturedHistory(jsonText)
}
\description{
\code{parseFeaturedHistory} parses the JSON returned by a featured history
request made to the appFigures web API.
}
