% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFeatured.R
\name{parseFeaturedSummary}
\alias{parseFeaturedSummary}
\title{Map JSON string to a list of R data frames.}
\usage{
parseFeaturedSummary(jsonText)
}
\description{
\code{parseFeaturedSummary} parses the JSON returned by a
featured summary request made to the appFigures web API.
}

