% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getStoreData.R
\name{parseFullStoreData}
\alias{parseFullStoreData}
\title{Map JSON string to an R data frame.}
\usage{
parseFullStoreData(jsonText)
}
\description{
\code{parseFullStoreData} parses the JSON returned by a
request made to the appFigures web API.
}

