% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProducts.R
\name{parseProducts}
\alias{parseProducts}
\title{Map JSON string to an R data frame.}
\usage{
parseProducts(jsonText)
}
\description{
\code{parseProducts} parses the JSON returned by a product
request made to the appFigures web API.
}

