% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{parseRanks}
\alias{parseRanks}
\title{Map JSON string to an R data frame.}
\usage{
parseRanks(jsonText)
}
\description{
\code{parseRanks} parses the JSON returned by an app rank request made to the
 appFigures web API.
}
