% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReviews.R
\name{parseReviewCounts}
\alias{parseReviewCounts}
\title{Map JSON string to a list of named vectors.}
\usage{
parseReviewCounts(jsonText)
}
\description{
\code{parseReviewCounts} parses the JSON returned by an app
review request made to the appFigures web API.
}
