% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReviews.R
\name{parseReviews}
\alias{parseReviews}
\title{Map JSON string to an R data frame.}
\usage{
parseReviews(jsonText)
}
\description{
\code{parseReviews} parses the JSON returned by an app
review request made to the appFigures web API.
}
