% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStoreData.R
\name{parseSDKData}
\alias{parseSDKData}
\title{Map JSON string to an R data frame.}
\usage{
parseSDKData(jsonText)
}
\description{
\code{parseSDKData} parses the JSON returned by a
request made to the appFigures web API.
}
