% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchProducts.R
\name{parseSearch}
\alias{parseSearch}
\title{Map JSON string to an R data frame.}
\usage{
parseSearch(jsonText)
}
\description{
\code{parseSearch} parses the JSON returned by a search
request made to the appFigures web API.
}

