% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRanks.R
\name{parseSnapshot}
\alias{parseSnapshot}
\title{Map JSON string to an R data frame.}
\usage{
parseSnapshot(jsonText)
}
\description{
\code{parseSnapshot} parses the JSON returned by a snapshot
request made to the appFigures web API.
}

