% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getUsage.R
\name{parseUsage}
\alias{parseUsage}
\title{Map JSON string to an R data frame.}
\usage{
parseUsage(jsonText)
}
\description{
\code{parseUsage} parses the JSON returned by a
usage request made to the appFigures web API.
}

