% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRanks.R
\name{stackSnapshots}
\alias{stackSnapshots}
\title{Retrieve multiple rank snapshots.}
\usage{
stackSnapshots(tsmps, curlHandle = NULL, verbose = F, ...)
}
\arguments{
\item{tsmps}{Either a character vector of elements with the
format: "yyyy-MM-ddTHH", or a vector of POSIX time stamps.
Only the last 24 hours are available.}

\item{curlHandle}{Provide an instance of the CURLHandle-class
from the RCurl package. The default will create a curl handle
specific to the function call.}

\item{verbose}{Logical. Should details of the web request
print to the console? Defaults to \code{FALSE}.}

\item{...}{Other arguments to pass to \code{\link{getRankSnapshot}}.
Currently, \code{stackSnapshots} is simple a wrapper to
\code{getRankSnapshot}.}
}
\value{
A data frame containing product and category data for
a cross section of top ranked apps.
}
\description{
\code{stackSnapshots} requests data for the top ranked apps for
for multiple time stamps from the appFigures web API.
}
\details{
See \code{\link{getRankSnapshot}}.
}
\seealso{
Official documentation:
\url{http://docs.appfigures.com/api/reference/v2/ranks}.
}

