% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFeatured.R
\name{parseFeaturedReport}
\alias{parseFeaturedReport}
\title{Map JSON string to an R data frame.}
\usage{
parseFeaturedReport(jsonText, pid)
}
\description{
\code{parseFeaturedReport} parses the JSON returned by a
featured request made to the appFigures web API.
}

